Change support field  to float64

+++ ruleserver/associationrules.go
@@ -6,7 +6,7 @@ import "fmt"
 type AssociationRule struct {
 	A          []string `json:"a"`
 	B          []string `json:"b"`
-	Support    int      `json:"support"`
+	Support    float64  `json:"support"`
 	Confidence float64  `json:"confidence"`
 	Lift       float64  `json:"lift"`
 }
@@ -15,5 +15,5 @@ type AssociationRule struct {
 type AssociationRules []AssociationRule
 
 func (a AssociationRule) String() string {
-	return fmt.Sprintf("%s => %s, sup: %d, conf: %f, lift: %f", a.A, a.B, a.Support, a.Confidence, a.Lift)
+	return fmt.Sprintf("%s => %s, sup: %f, conf: %f, lift: %f", a.A, a.B, a.Support, a.Confidence, a.Lift)
 }
